Move base test cases to their own dir & export them
The directory search component extends the InputTestCase, which is how I intended for folks to do testing on their stuff.

Moving it into its own directory lets me expose it to the autoloader without exposing too much.
