+++ tests/Components/Layout/ContentTest.php
@@ -3,7 +3,7 @@
 namespace Northwestern\SysDev\DynamicForms\Tests\Components\Layout;
 
 use Northwestern\SysDev\DynamicForms\Components\Layout\Content;
-use Northwestern\SysDev\DynamicForms\Tests\Components\BaseComponentTestCase;
+use Northwestern\SysDev\DynamicForms\Tests\Components\TestCases\BaseComponentTestCase;
 
 /**
  * @coversDefaultClass \Northwestern\SysDev\DynamicForms\Components\Layout\Content
