+++ tests/Components/TestCases/InputComponentTestCase.php
@@ -1,11 +1,14 @@
 <?php
 
-namespace Northwestern\SysDev\DynamicForms\Tests\Components;
+namespace Northwestern\SysDev\DynamicForms\Tests\Components\TestCases;
+
+use Northwestern\SysDev\DynamicForms\Tests\Components\TestCases\BaseComponentTestCase;
+use function app;
 
 /**
  * @coversDefaultClass \Northwestern\SysDev\DynamicForms\Components\BaseComponent
  */
-class InputComponentTestCase extends BaseComponentTestCase
+abstract class InputComponentTestCase extends BaseComponentTestCase
 {
     /**
      * @covers ::canValidate
